+++ packages/pagination/src/elements/CursorPagination/components/Last.tsx
@@ -5,11 +5,11 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import React, { HTMLAttributes } from 'react';
+import React, { ButtonHTMLAttributes } from 'react';
 import ChevronDoubleRight from '@zendeskgarden/svg-icons/src/16/chevron-double-right-stroke.svg';
 import { StyledIcon, StyledCursor } from '../../../styled';
 
-export const Last = React.forwardRef<HTMLButtonElement, HTMLAttributes<HTMLButtonElement>>(
+export const Last = React.forwardRef<HTMLButtonElement, ButtonHTMLAttributes<HTMLButtonElement>>(
   ({ children, ...other }, ref) => {
     return (
       <StyledCursor ref={ref} as="button" {...other}>